refactor: separate tags from main
